refactor(History): make the class concrete.
